Generalize Targets as Operators [pass 3]
